refactor: remove daemon/handlers.ts re-export shim
