+++ assistant/src/__tests__/handlers-add-trust-rule-metadata.test.ts
@@ -65,8 +65,8 @@ mock.module("../config/loader.js", () => ({
   setNestedValue: () => {},
 }));
 
-import type { HandlerContext } from "../daemon/handlers.js";
 import { handleAddTrustRule } from "../daemon/handlers/config.js";
+import type { HandlerContext } from "../daemon/handlers/shared.js";
 import type { AddTrustRule } from "../daemon/ipc-contract.js";
 import type { ServerMessage } from "../daemon/ipc-contract.js";
 import {
