+++ assistant/src/__tests__/handlers-cu-observation-blob.test.ts
@@ -49,7 +49,8 @@ mock.module("../util/logger.js", () => ({
 }));
 
 import { ComputerUseSession } from "../daemon/computer-use-session.js";
-import { handleMessage, type HandlerContext } from "../daemon/handlers.js";
+import { handleMessage } from "../daemon/handlers/index.js";
+import type { HandlerContext } from "../daemon/handlers/shared.js";
 import type {
   CuObservation,
   IpcBlobRef,
