+++ assistant/src/__tests__/handlers-ipc-blob-probe.test.ts
@@ -42,7 +42,8 @@ mock.module("../util/logger.js", () => ({
   }),
 }));
 
-import { handleMessage, type HandlerContext } from "../daemon/handlers.js";
+import { handleMessage } from "../daemon/handlers/index.js";
+import type { HandlerContext } from "../daemon/handlers/shared.js";
 import type { IpcBlobProbe, ServerMessage } from "../daemon/ipc-contract.js";
 import { DebouncerMap } from "../util/debounce.js";
