+++ assistant/src/__tests__/handlers-user-message-approval-consumption.test.ts
@@ -3,7 +3,7 @@ import { afterAll, beforeEach, describe, expect, mock, test } from "bun:test";
 
 mock.module("../config/env.js", () => ({ isHttpAuthDisabled: () => true }));
 
-import type { HandlerContext } from "../daemon/handlers.js";
+import type { HandlerContext } from "../daemon/handlers/shared.js";
 import type {
   ConfirmationResponse,
   UserMessage,
