+++ assistant/src/__tests__/tool-permission-simulate-handler.test.ts
@@ -61,8 +61,8 @@ mock.module("../config/loader.js", () => ({
   setNestedValue: () => {},
 }));
 
-import type { HandlerContext } from "../daemon/handlers.js";
 import { handleToolPermissionSimulate } from "../daemon/handlers/config.js";
+import type { HandlerContext } from "../daemon/handlers/shared.js";
 import type {
   ServerMessage,
   ToolPermissionSimulateRequest,
