+++ assistant/src/daemon/ride-shotgun-handler.ts
@@ -20,7 +20,7 @@ import {
   watchSessions,
 } from "../tools/watch/watch-state.js";
 import { getLogger } from "../util/logger.js";
-import type { HandlerContext } from "./handlers.js";
+import type { HandlerContext } from "./handlers/shared.js";
 import type { RideShotgunStart, RideShotgunStop } from "./ipc-protocol.js";
 import { generateSummary, lastSummaryBySession } from "./watch-handler.js";
